make that work on windows

+++ IPlugExamples/IPlugEffect/IPlugEffectControls.h
@@ -93,29 +93,29 @@ public:
   {
     graphics.FillRect(COLOR_GRAY, mRECT.GetPadded(-2));
     graphics.DrawRect(COLOR_BLACK, mRECT.GetPadded(-2));
-    double angle = mAngle1 + mValue * (mAngle2 - mAngle1);
-    graphics.FillArc(COLOR_BLUE, mRECT.MW(), mRECT.MH(), mRECT.W() * 0.44, mAngle1, angle);
-    graphics.DrawArc(COLOR_BLACK, mRECT.MW(), mRECT.MH(), mRECT.W() * 0.44, mAngle1, angle);
-    graphics.DrawRadialLine(COLOR_BLACK, mRECT.MW(), mRECT.MH(), angle, 0., mRECT.W() * 0.49);
-    graphics.FillCircle(COLOR_WHITE, mRECT.MW(), mRECT.MH(), mRECT.W() * 0.1);
-    graphics.DrawCircle(COLOR_BLACK, mRECT.MW(), mRECT.MH(), mRECT.W() * 0.1);
+    float angle = mAngle1 + (float) mValue * (mAngle2 - mAngle1);
+    graphics.FillArc(COLOR_BLUE, mRECT.MW(), mRECT.MH(), mRECT.W() * 0.44f, mAngle1, angle);
+    graphics.DrawArc(COLOR_BLACK, mRECT.MW(), mRECT.MH(), mRECT.W() * 0.44f, mAngle1, angle);
+    graphics.DrawRadialLine(COLOR_BLACK, mRECT.MW(), mRECT.MH(), angle, 0.f, mRECT.W() * 0.49f);
+    graphics.FillCircle(COLOR_WHITE, mRECT.MW(), mRECT.MH(), mRECT.W() * 0.1f);
+    graphics.DrawCircle(COLOR_BLACK, mRECT.MW(), mRECT.MH(), mRECT.W() * 0.1f);
     
     angle = DegToRad(angle);
     
-    float x1 = mRECT.MW() + cos(angle - 0.3) * mRECT.W() * 0.3;
-    float y1 = mRECT.MH() + sin(angle - 0.3) * mRECT.W() * 0.3;
-    float x2 = mRECT.MW() + cos(angle + 0.3) * mRECT.W() * 0.3;
-    float y2 = mRECT.MH() + sin(angle + 0.3) * mRECT.W() * 0.3;
-    float x3 = mRECT.MW() + cos(angle) * mRECT.W() * 0.44;
-    float y3 = mRECT.MH() + sin(angle) * mRECT.W() * 0.44;
+    float x1 = mRECT.MW() + cosf(angle - 0.3f) * mRECT.W() * 0.3f;
+    float y1 = mRECT.MH() + sinf(angle - 0.3f) * mRECT.W() * 0.3f;
+    float x2 = mRECT.MW() + cosf(angle + 0.3f) * mRECT.W() * 0.3f;
+    float y2 = mRECT.MH() + sinf(angle + 0.3f) * mRECT.W() * 0.3f;
+    float x3 = mRECT.MW() + cosf(angle) * mRECT.W() * 0.44f;
+    float y3 = mRECT.MH() + sinf(angle) * mRECT.W() * 0.44f;
     
     graphics.FillTriangle(COLOR_WHITE, x1, y1, x2, y2, x3, y3);
     graphics.DrawTriangle(COLOR_BLACK, x1, y1, x2, y2, x3, y3);
   }
   
 private:
-  double mAngle1;
-  double mAngle2;
+  float mAngle1;
+  float mAngle2;
 };
 
 class IPolyControl : public IKnobControlBase
@@ -131,18 +131,18 @@ private:
   {
     float xarray[32];
     float yarray[32];
-    int npoints = 3 + round(mValue * 29);
-    double angle = (-0.75 * PI) + mValue * (1.5 * PI);
-    double incr = (2 * PI) / npoints;
-    double cr = mValue * (mRECT.W() / 2.0);
+    int npoints = 3 + (int) roundf((float) mValue * 29.f);
+    float angle = (-0.75f * (float) PI) + (float) mValue * (1.5f * (float) PI);
+    float incr = (2.f * (float) PI) / npoints;
+    float cr = (float) mValue * (mRECT.W() / 2.f);
     
-    graphics.FillRoundRect(COLOR_GRAY, mRECT.GetPadded(-2), cr);
-    graphics.DrawRoundRect(COLOR_BLACK, mRECT.GetPadded(-2), cr);
+    graphics.FillRoundRect(COLOR_GRAY, mRECT.GetPadded(-2.f), cr);
+    graphics.DrawRoundRect(COLOR_BLACK, mRECT.GetPadded(-2.f), cr);
     
     for (int i = 0; i < npoints; i++)
     {
-      xarray[i] = mRECT.MW() + sin(angle + i * incr) * mRECT.W() * 0.45;
-      yarray[i] = mRECT.MH() + cos(angle + i * incr) * mRECT.W() * 0.45;
+      xarray[i] = mRECT.MW() + sinf(angle + (float) i * incr) * mRECT.W() * 0.45f;
+      yarray[i] = mRECT.MH() + cosf(angle + (float) i * incr) * mRECT.W() * 0.45f;
     }
     
     graphics.FillConvexPolygon(COLOR_ORANGE, xarray, yarray, npoints);
